feat: follow golang project structure + update to go 1.24
