+++ internal/constants.go
@@ -1,0 +1,6 @@
+package internal
+
+const BASE = 62
+const CODE_LENGTH = 7
+
+var CHARACTERS = [62]string{"p", "X", "4", "k", "V", "b", "N", "8", "i", "Q", "w", "Z", "2", "F", "m", "Y", "j", "U", "9", "R", "d", "H", "5", "L", "n", "C", "a", "W", "1", "G", "y", "D", "e", "0", "S", "h", "B", "7", "t", "O", "c", "I", "v", "3", "M", "q", "T", "r", "6", "E", "K", "f", "A", "x", "J", "g", "P", "s", "l", "u", "o", "z"}
